# cmd/secretsbroker/main.go
package main

import (
	"context"
	"encoding/json"
	"errors"
	"flag"
	"fmt"
	"log/slog"
	"net/http"
	"os"
	"os/signal"
	"strings"
	"syscall"
	"time"
)

const (
	version         = "0.1.0"
	apiVersion      = "secretsbroker.local/v1"
	contractVersion = "1.1.0"
	serviceID       = "@secretsbroker"
)

var lifecycleStates = []string{
	"setup_needed",
	"locked",
	"ready",
	"source_auth_required",
	"degraded",
	"policy_denied",
}

var typedOutcomes = []string{
	"setup_needed",
	"locked",
	"ready",
	"source_auth_required",
	"degraded",
	"policy_denied",
	"missing_ref",
	"invalid_ref",
	"source_unavailable",
	"conflict",
	"staged",
	"applied",
	"rolled_back",
	"retired",
	"identity_invalid",
	"identity_expired",
	"identity_replayed",
	"lockout_active",
	"cleared",
	"not_found",
	"disabled",
	"stale",
	"unsupported",
	"audit_unavailable",
	"dependency_blocked",
	"stale_plan",
	"failed",
}

type Status struct {
	ServiceID   string    `json:"serviceId"`
	Name        string    `json:"name"`
	Version     string    `json:"version"`
	APIVersion  string    `json:"apiVersion"`
	State       string    `json:"state"`
	Ready       bool      `json:"ready"`
	LocalFirst  bool      `json:"localFirst"`
	Backend     string    `json:"backend"`
	Description string    `json:"description"`
	CheckedAt   time.Time `json:"checkedAt"`
}

type HealthResponse struct {
	OK        bool   `json:"ok"`
	ServiceID string `json:"serviceId"`
	State     string `json:"state"`
}

type StateResponse struct {
	ServiceID        string   `json:"serviceId"`
	APIVersion       string   `json:"apiVersion"`
	State            string   `json:"state"`
	Ready            bool     `json:"ready"`
	Outcome          string   `json:"outcome"`
	KeyState         string   `json:"keyState"`
	NextAction       string   `json:"nextAction"`
	AffectedRefs     []string `json:"affectedRefs"`
	AffectedServices []string `json:"affectedServices"`
}

type CapabilitiesResponse struct {
	ServiceID       string                `json:"serviceId"`
	APIVersion      string                `json:"apiVersion"`
	ContractVersion string                `json:"contractVersion"`
	ManifestVersion string                `json:"manifestVersion"`
	Version         string                `json:"version"`
	Transports      []string              `json:"transports"`
	Endpoints       []string              `json:"endpoints"`
	Operations      []OperationCapability `json:"operations"`
	Features        []string              `json:"features"`
	FutureFeatures  []string              `json:"futureFeatures"`
	Outcomes        []string              `json:"outcomes"`
}

type ErrorEnvelope struct {
	Error APIError `json:"error"`
}

type APIError struct {
	Code              string   `json:"code"`
	Message           string   `json:"message"`
	Outcome           string   `json:"outcome"`
	RequestID         string   `json:"requestId,omitempty"`
	NextAction        string   `json:"nextAction,omitempty"`
	AffectedRefs      []string `json:"affectedRefs"`
	AffectedServices  []string `json:"affectedServices"`
	LockoutActive     bool     `json:"lockoutActive,omitempty"`
	LockoutScope      string   `json:"lockoutScope,omitempty"`
	RetryAfterSeconds int      `json:"retryAfterSeconds,omitempty"`
}

func main() {
	if err := run(os.Args[1:]); err != nil {
		fmt.Fprintln(os.Stderr, err)
		os.Exit(1)
	}
}

func run(args []string) error {
	if len(args) == 0 {
		args = []string{"serve"}
	}

	switch args[0] {
	case "serve":
		return serve(args[1:])
	case "status":
		return printStatus(args[1:])
	case "key":
		return runKey(args[1:])
	case "backup":
		return runBackup(args[1:])
	case "admin":
		return runAdmin(args[1:])
	case "session":
		return runSession(args[1:])
	case "version":
		fmt.Println(version)
		return nil
	case "help", "-h", "--help":
		usage()
		return nil
	default:
		usage()
		return fmt.Errorf("unknown command %q", args[0])
	}
}

func usage() {
	fmt.Fprintln(os.Stderr, "Usage: secretsbroker <serve|status|key|backup|admin|session|version>")
	fmt.Fprintln(os.Stderr, "")
	fmt.Fprintln(os.Stderr, "Commands:")
	fmt.Fprintln(os.Stderr, "  serve   Start the local-first @secretsbroker daemon")
	fmt.Fprintln(os.Stderr, "  status  Print current broker state JSON")
	fmt.Fprintln(os.Stderr, "  key     Manage portable master-key lifecycle, local wrapper, and rotation")
	fmt.Fprintln(os.Stderr, "  backup  Create or restore encrypted local-store backup artifacts")
	fmt.Fprintln(os.Stderr, "  admin   Headless admin CLI for status, metadata search, reveal, provider, migration, and audit workflows")
	fmt.Fprintln(os.Stderr, "  session Manage local API session/token foundation")
	fmt.Fprintln(os.Stderr, "  version Print broker version")
}

func normalizeState(state string) string {
	state = strings.TrimSpace(state)
	if state == "" {
		return "setup_needed"
	}
	for _, allowed := range lifecycleStates {
		if state == allowed {
			return state
		}
	}
	return "degraded"
}

func isReadyState(state string) bool {
	return normalizeState(state) == "ready"
}

func defaultStatus(state string) Status {
	state = normalizeState(state)
	return Status{
		ServiceID:   serviceID,
		Name:        "Service Lasso Secrets Broker",
		Version:     version,
		APIVersion:  apiVersion,
		State:       state,
		Ready:       isReadyState(state),
		LocalFirst:  true,
		Backend:     "local",
		Description: "Local-first secrets broker with encrypted local storage, metadata-safe management, resolve, write-back, and provisioning status surfaces.",
		CheckedAt:   time.Now().UTC(),
	}
}

func defaultHealth(state string) HealthResponse {
	return HealthResponse{OK: true, ServiceID: serviceID, State: normalizeState(state)}
}

func defaultState(state string) StateResponse {
	return stateResponse(state, nil, nil)
}

func stateResponse(state string, affectedRefs []string, affectedServices []string) StateResponse {
	state = normalizeState(state)
	return StateResponse{
		ServiceID:        serviceID,
		APIVersion:       apiVersion,
		State:            state,
		Ready:            isReadyState(state),
		Outcome:          state,
		KeyState:         keyStateFor(state),
		NextAction:       nextActionFor(state),
		AffectedRefs:     safeList(affectedRefs),
		AffectedServices: safeList(affectedServices),
	}
}

func defaultCapabilities() CapabilitiesResponse {
	return CapabilitiesResponse{
		ServiceID:       serviceID,
		APIVersion:      apiVersion,
		ContractVersion: contractVersion,
		ManifestVersion: operationManifestVersion,
		Version:         version,
		Transports:      []string{"loopback-http", "unix-socket", "windows-named-pipe"},
		Endpoints: []string{
			"GET /health",
			"GET /ready",
			"GET /status",
			"GET /state",
			"GET /capabilities",
			"POST /v1/secrets",
			"POST /v1/writeback",
			"POST /v1/resolve",
			"GET /v1/provisioning/status",
			"POST /v1/provisioning/operations/plan",
			"POST /v1/provisioning/operations/apply",
			"GET /v1/sources/status",
			"GET /v1/providers/capabilities",
			"GET /v1/providers/config/status",
			"GET /v1/telemetry",
			"POST /v1/telemetry/export",
			"GET /v1/events",
			"GET|POST /v1/recovery/policy",
			"POST /v1/management/lockouts/clear",
			"POST /v1/providers/config/validate|apply",
			"POST /v1/providers/migration/dry-run|apply",
			"GET /v1/management/secrets",
			"GET /v1/management/secrets/value-search",
			"POST /v1/management/secrets/reveal",
			"POST /v1/management/secrets/edit/dry-run|apply",
			"POST /v1/management/secrets/reset/dry-run|apply",
			"POST /v1/management/secrets/decommission/dry-run|apply|restore",
			"POST /v1/management/secrets/rotation/dry-run",
			"POST /v1/management/secrets/rotation/status|stage|activate|rollback|retire",
			"POST /v1/management/secrets/campaigns/create|revalidate|apply|status",
			"POST /v1/management/secrets/sync/dry-run",
			"POST /v1/management/secrets/policy/preview|apply",
			"CLI secretsbroker admin mcp tools|call",
			"CLI secretsbroker backup create|restore",
			"CLI secretsbroker key initialize|unlock|import|rewrap|wrapper-status|rotate|recovery generate|recovery import",
			"CLI secretsbroker admin status|secrets|providers|migration|sync|recovery|audit|telemetry|events",
			"CLI secretsbroker admin launch-lease issue",
		},
		Operations: defaultOperationManifest(),
		Features: []string{
			"liveness",
			"readiness",
			"status",
			"state",
			"capabilities",
			"versioned-operation-capability-manifest",
			"local-encrypted-store",
			"batched-resolve",
			"typed-errors",
			"audit-redaction",
			"source-status",
			"provider-config-status",
			"provider-config-validation",
			"redacted-telemetry",
			"otel-telemetry-preview",
			"otel-telemetry-export-action",
			"bounded-operational-events",
			"scoped-local-api-lockout",
			"audited-lockout-clear",
			"provider-migration-dry-run-apply",
			"secrets-management-metadata-search",
			"secrets-management-value-search-metadata-only",
			"secrets-management-controlled-reveal",
			"secrets-management-dry-run-apply",
			"local-secret-decommission-tombstone-restore",
			"credential-rotation-dry-run",
			"credential-rotation-local-versioning",
			"bulk-secret-operation-campaigns",
			"metadata-only-secrets-sync-dry-run",
			"env-source",
			"file-source",
			"exec-source",
			"service-json-secret-policy",
			"signed-launch-identity-leases",
			"launch-identity-transport-binding",
			"write-back-policy",
			"generated-secret-capture",
			"generated-secret-provisioning-status",
			"generated-secret-provisioning-operation-plan",
			"generated-secret-provisioning-broker-apply",
			"encrypted-backup-restore",
			"master-key-initialize-unlock-import-rewrap",
			"os-wrapper-status",
			"master-key-rotation",
			"recovery-policy-metadata",
			"recovery-policy-status",
			"threshold-recovery-shares",
			"headless-admin-cli",
			"metadata-only-mcp-adapter",
			"os-ipc-transport-policy",
			"unix-socket-peer-credential-checks",
			"windows-named-pipe-identity-checks",
			"windows-named-pipe-service-account-policy",
			"launcher-issued-transport-bound-leases",
		},
		FutureFeatures: []string{
			"external-backend-write-back",
		},
		Outcomes: append([]string(nil), typedOutcomes...),
	}
}

func printStatus(args []string) error {
	fs := flag.NewFlagSet("status", flag.ContinueOnError)
	state := fs.String("state", getenvDefault("SECRETSBROKER_STATE", "setup_needed"), "state to report")
	if err := fs.Parse(args); err != nil {
		return err
	}
	enc := json.NewEncoder(os.Stdout)
	enc.SetIndent("", "  ")
	return enc.Encode(defaultStatus(*state))
}

func serve(args []string) error {
	fs := flag.NewFlagSet("serve", flag.ContinueOnError)
	listen := fs.String("listen", getenvDefault("SECRETSBROKER_LISTEN", "127.0.0.1:17890"), "listen address")
	mode := fs.String("mode", getenvDefault("SECRETSBROKER_MODE", "development"), "runtime mode: development or production")
	transport := fs.String("transport", getenvDefault("SECRETSBROKER_TRANSPORT", "loopback-http"), "serve transport: loopback-http, unix-socket, windows-named-pipe, or auto")
	unixSocket := fs.String("unix-socket", getenvDefault("SECRETSBROKER_UNIX_SOCKET", ""), "Unix socket path for unix-socket transport")
	namedPipe := fs.String("named-pipe", getenvDefault("SECRETSBROKER_NAMED_PIPE", ""), "Windows named pipe path for windows-named-pipe transport")
	namedPipeAllowedSIDs := multiFlag(splitCSV(getenvDefault("SECRETSBROKER_NAMED_PIPE_ALLOWED_SIDS", "")))
	namedPipeAllowAdmin := fs.Bool("named-pipe-allow-admin", envBoolDefault("SECRETSBROKER_NAMED_PIPE_ALLOW_ADMIN", true), "allow enabled local Administrators group members to connect to the Windows named-pipe transport")
	namedPipeAllowLocalSystem := fs.Bool("named-pipe-allow-local-system", envBoolDefault("SECRETSBROKER_NAMED_PIPE_ALLOW_LOCAL_SYSTEM", true), "allow LocalSystem to connect to the Windows named-pipe transport")
	state := fs.String("state", getenvDefault("SECRETSBROKER_STATE", "setup_needed"), "state to report")
	storePath := fs.String("store", getenvDefault("SECRETSBROKER_STORE_PATH", defaultStorePath()), "local encrypted store path")
	auditPath := fs.String("audit", getenvDefault("SECRETSBROKER_AUDIT_PATH", defaultAuditPath()), "audit JSONL path")
	eventsPath := fs.String("events", getenvDefault("SECRETSBROKER_EVENTS_PATH", defaultEventsPath(*auditPath)), "operational events JSONL path")
	auditHashChain := fs.Bool("audit-hash-chain", envBoolDefault("SECRETSBROKER_AUDIT_HASH_CHAIN", false), "append tamper-evident audit hash-chain metadata")
	masterKey := fs.String("master-key", getenvDefault("SECRETSBROKER_MASTER_KEY", ""), "local development master key; empty means locked")
	masterKeyFile := fs.String("master-key-file", getenvDefault("SECRETSBROKER_MASTER_KEY_FILE", ""), "file containing portable master key")
	apiToken := fs.String("api-token", getenvDefault("SECRETSBROKER_API_TOKEN", ""), "local API token for secret-bearing endpoints")
	launchIdentitySigningKey := fs.String("launch-identity-signing-key", getenvDefault("SECRETSBROKER_LAUNCH_IDENTITY_SIGNING_KEY", ""), "HMAC key for signed scoped launch identity leases; defaults to local API token in bootstrap mode")
	sourcesPath := fs.String("sources", getenvDefault("SECRETSBROKER_SOURCES_PATH", ""), "source adapter config path")
	affectedRefs := multiFlag(splitCSV(getenvDefault("SECRETSBROKER_AFFECTED_REFS", "")))
	affectedServices := multiFlag(splitCSV(getenvDefault("SECRETSBROKER_AFFECTED_SERVICES", "")))
	fs.Var(&namedPipeAllowedSIDs, "named-pipe-allowed-sid", "additional Windows service-account/user SID allowed to connect to the named-pipe transport; repeatable")
	fs.Var(&affectedRefs, "affected-ref", "affected secret ref to report for non-ready states; repeatable")
	fs.Var(&affectedServices, "affected-service", "affected service id to report for non-ready states; repeatable")
	if err := fs.Parse(args); err != nil {
		return err
	}
	eventsPathValue := *eventsPath
	eventsPathExplicit := false
	fs.Visit(func(f *flag.Flag) {
		if f.Name == "events" {
			eventsPathExplicit = true
		}
	})
	if !eventsPathExplicit && strings.TrimSpace(os.Getenv("SECRETSBROKER_EVENTS_PATH")) == "" {
		eventsPathValue = defaultEventsPath(*auditPath)
	}

	refs := []string(affectedRefs)
	services := []string(affectedServices)
	stateView := runtimeState{state: state, affectedRefs: &refs, affectedServices: &services}
	material, err := loadKeyMaterial(*masterKey, *masterKeyFile)
	if err != nil && !errors.Is(err, errLocked) {
		return err
	}
	backend := newLocalBackend(*storePath, *auditPath, material.Value)
	backend.eventPath = eventsPathValue
	backend.auditHashChain = *auditHashChain
	backend.launchIdentitySigningKey = strings.TrimSpace(*launchIdentitySigningKey)
	sources, err := loadSourceConfig(*sourcesPath)
	if err != nil {
		return err
	}
	backend.sources = sources

	binding, err := resolveServeTransport(serveTransportOptions{
		Mode:                        *mode,
		Transport:                   *transport,
		Listen:                      *listen,
		UnixSocket:                  *unixSocket,
		NamedPipe:                   *namedPipe,
		NamedPipeAllowedSIDs:        []string(namedPipeAllowedSIDs),
		NamedPipeAllowLocalSystem:   *namedPipeAllowLocalSystem,
		NamedPipeAllowBuiltinAdmins: *namedPipeAllowAdmin,
	})
	if err != nil {
		return err
	}
	ln, cleanup, err := listenForTransport(binding)
	if err != nil {
		return err
	}
	defer cleanup()

	server := &http.Server{
		Handler:           newHandler(stateView, backend, localAPISecurity{token: *apiToken}),
		ReadHeaderTimeout: 5 * time.Second,
		ConnContext:       transportPeerIdentityConnContext,
	}
	go func() {
		slog.Info("@secretsbroker listening", "transport", binding.Kind, "addr", binding.DisplayAddress, "state", *state)
		if err := server.Serve(ln); err != nil && !errors.Is(err, http.ErrServerClosed) {
			slog.Error("server failed", "error", err)
		}
	}()

	ctx, stop := signal.NotifyContext(context.Background(), os.Interrupt, syscall.SIGTERM)
	defer stop()
	<-ctx.Done()

	shutdownCtx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
	defer cancel()
	return server.Shutdown(shutdownCtx)
}

func newHandler(state runtimeState, backend *localBackend, security localAPISecurity) http.Handler {
	mux := http.NewServeMux()
	mux.HandleFunc("/health", func(w http.ResponseWriter, r *http.Request) {
		writeJSON(w, http.StatusOK, defaultHealth(state.current()))
	})
	mux.HandleFunc("/ready", func(w http.ResponseWriter, r *http.Request) {
		body := state.response()
		code := http.StatusOK
		if !body.Ready {
			code = http.StatusServiceUnavailable
		}
		writeJSON(w, code, body)
	})
	mux.HandleFunc("/status", func(w http.ResponseWriter, r *http.Request) {
		writeJSON(w, http.StatusOK, defaultStatus(state.current()))
	})
	mux.HandleFunc("/state", func(w http.ResponseWriter, r *http.Request) {
		writeJSON(w, http.StatusOK, state.response())
	})
	mux.HandleFunc("/capabilities", func(w http.ResponseWriter, r *http.Request) {
		writeJSON(w, http.StatusOK, defaultCapabilities())
	})
	if backend != nil {
		if security.lockouts == nil {
			security.lockouts = newLockoutStore(nil)
		}
		if security.audit == nil {
			security.audit = backend.audit
		}
		registerLocalStoreHandlers(mux, backend, security)
		registerSourceRegistryHandlers(mux, backend)
		registerSecretsManagementHandlers(mux, backend, security)
		registerDecommissionHandlers(mux, backend, security)
		registerRotationHandlers(mux, backend, security)
		registerRotationVersionHandlers(mux, backend, security)
		registerBulkCampaignHandlers(mux, backend, security)
		registerSyncDryRunHandlers(mux, backend, security)
		registerProviderConfigMigrationHandlers(mux, backend, security)
		registerRecoveryPolicyHandlers(mux, backend, security)
		registerTelemetryHandlers(mux, backend)
		registerEventsHandlers(mux, backend)
		registerLockoutManagementHandlers(mux, backend, security)
	}
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		started := time.Now()
		applyTelemetryResponseHeaders(w, r)
		recorder := &telemetryStatusRecorder{ResponseWriter: w, statusCode: http.StatusOK}
		mux.ServeHTTP(recorder, r)
		if backend != nil {
			recordTelemetryHTTPRequest(backend.telemetry, r, recorder.statusCode, time.Since(started))
		}
	})
}

type runtimeState struct {
	state            *string
	affectedRefs     *[]string
	affectedServices *[]string
}

func (s runtimeState) current() string {
	if s.state == nil {
		return "setup_needed"
	}
	return *s.state
}

func (s runtimeState) response() StateResponse {
	var refs []string
	if s.affectedRefs != nil {
		refs = *s.affectedRefs
	}
	var services []string
	if s.affectedServices != nil {
		services = *s.affectedServices
	}
	return stateResponse(s.current(), refs, services)
}

type multiFlag []string

func (m *multiFlag) String() string { return strings.Join(*m, ",") }

func (m *multiFlag) Set(value string) error {
	*m = append(*m, splitCSV(value)...)
	return nil
}

func splitCSV(value string) []string {
	if strings.TrimSpace(value) == "" {
		return nil
	}
	parts := strings.Split(value, ",")
	return safeList(parts)
}

func safeList(values []string) []string {
	if len(values) == 0 {
		return []string{}
	}
	clean := make([]string, 0, len(values))
	for _, value := range values {
		value = strings.TrimSpace(value)
		if value != "" {
			clean = append(clean, value)
		}
	}
	return clean
}

func keyStateFor(state string) string {
	switch normalizeState(state) {
	case "setup_needed":
		return "not_initialized"
	case "locked":
		return "locked"
	default:
		return "available"
	}
}

func nextActionFor(state string) string {
	switch normalizeState(state) {
	case "setup_needed":
		return "run_setup"
	case "locked":
		return "unlock_broker"
	case "source_auth_required":
		return "reconnect_source"
	case "degraded":
		return "inspect_sources"
	case "policy_denied":
		return "review_policy"
	default:
		return ""
	}
}

func writeJSON(w http.ResponseWriter, code int, value any) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(code)
	_ = json.NewEncoder(w).Encode(value)
}

func getenvDefault(key, fallback string) string {
	if value := os.Getenv(key); value != "" {
		return value
	}
	return fallback
}

func envBoolDefault(key string, fallback bool) bool {
	switch strings.ToLower(strings.TrimSpace(os.Getenv(key))) {
	case "1", "true", "yes", "on":
		return true
	case "0", "false", "no", "off":
		return false
	default:
		return fallback
	}
}