# contract/v1/openapi.json
{
  "components": {
    "schemas": {
...
      "CapabilitiesResponse": {
        "additionalProperties": true,
        "properties": {
          "apiVersion": {
            "type": "string"
          },
          "contractVersion": {
            "type": "string"
          },
          "endpoints": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "features": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "futureFeatures": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "manifestVersion": {
            "type": "string"
          },
          "operations": {
            "items": {
              "$ref": "#/components/schemas/OperationCapability"
            },
            "type": "array"
          },
          "outcomes": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "serviceId": {
            "type": "string"
          },
          "transports": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "version": {
            "type": "string"
          }
        },
        "required": [
          "apiVersion",
          "contractVersion",
          "endpoints",
          "features",
          "futureFeatures",
          "manifestVersion",
          "operations",
          "outcomes",
          "serviceId",
          "transports",
          "version"
        ],
        "type": "object"
      },
      "ErrorEnvelope": {
        "additionalProperties": true,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/APIError"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "localBearerToken": {
        "scheme": "bearer",
        "type": "http"
      },
      "localHeaderToken": {
        "in": "header",
        "name": "X-SecretsBroker-Token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Canonical metadata-safe contract generated from the Go request and response DTOs. Unknown response fields are compatible additions.",
    "title": "Service Lasso Secrets Broker local API",
    "version": "1.1.0"
  },
...
  "openapi": "3.1.0",
  "paths": {
    "/capabilities": {
      "get": {
        "operationId": "getcapabilities",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapabilitiesResponse"
                }
              }
            },
            "description": "Successful response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/CapabilitiesResponse"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope"
                    }
                  ]
                }
              }
            },
            "description": "Typed failure response"
          }
        },
        "security": [],
        "summary": "Broker routes, features and outcomes",
        "x-capability": {
          "operationId": "getcapabilities",
          "method": "GET",
          "path": "/capabilities",
          "maturity": "read-only",
          "classification": "read",
          "authenticationRequired": false,
          "policyRequired": false,
          "auditRequired": false,
          "scope": "broker-local",
          "providerKinds": [],
          "completionMode": "synchronous",
          "limitationCode": "release_metadata_only",
          "reasonCode": "implemented_read_only",
          "nextAction": "read_when_authorized"
        }
      }
    },
...
    "/state": {
      "get": {
        "operationId": "getstate",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StateResponse"
                }
              }
            },
            "description": "Successful response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/StateResponse"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope"
                    }
                  ]
                }
              }
            },
            "description": "Typed failure response"
          }
        },
        "security": [],
        "summary": "Typed broker lifecycle state",
        "x-capability": {
          "operationId": "getstate",
          "method": "GET",
          "path": "/state",
          "maturity": "read-only",
          "classification": "read",
          "authenticationRequired": false,
          "policyRequired": false,
          "auditRequired": false,
          "scope": "broker-local",
          "providerKinds": [],
          "completionMode": "synchronous",
          "limitationCode": "metadata_only",
          "reasonCode": "implemented_read_only",
          "nextAction": "read_when_authorized"
        }
      }
    },
    "/status": {
      "get": {
        "operationId": "getstatus",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            },
            "description": "Successful response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/Status"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope"
                    }
                  ]
                }
              }
            },
            "description": "Typed failure response"
          }
        },
        "security": [],
        "summary": "Safe broker status",
        "x-capability": {
          "operationId": "getstatus",
          "method": "GET",
          "path": "/status",
          "maturity": "read-only",
          "classification": "read",
          "authenticationRequired": false,
          "policyRequired": false,
          "auditRequired": false,
          "scope": "broker-local",
          "providerKinds": [],
          "completionMode": "synchronous",
          "limitationCode": "metadata_only",
          "reasonCode": "implemented_read_only",
          "nextAction": "read_when_authorized"
        }
      }
    },
    "/v1/events": {
      "get": {
        "operationId": "getv1_events",
        "parameters": [
          {
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "until",
            "required": false,
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "serviceId",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "providerId",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sourceId",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "operation",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "outcome",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "severity",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "family",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "refPrefix",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "refHash",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Maximum 200",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/eventsResponse"
                }
              }
            },
            "description": "Successful response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/eventsResponse"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope"
                    }
                  ]
                }
              }
            },
            "description": "Typed failure response"
          }
        },
        "security": [],
        "summary": "List bounded operational events",
        "x-capability": {
          "operationId": "getv1_events",
          "method": "GET",
          "path": "/v1/events",
          "maturity": "read-only",
          "classification": "read",
          "authenticationRequired": false,
          "policyRequired": false,
          "auditRequired": false,
          "scope": "broker-local",
          "providerKinds": [],
          "completionMode": "synchronous",
          "limitationCode": "metadata_only",
          "reasonCode": "implemented_read_only",
          "nextAction": "read_when_authorized"
        }
      }
    },
    "/v1/management/lockouts/clear": {
      "post": {
        "operationId": "postv1_management_lockouts_clear",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/lockoutClearRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/lockoutClearResponse"
                }
              }
            },
            "description": "Successful response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/lockoutClearResponse"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope"
                    }
                  ]
                }
              }
            },
            "description": "Typed failure response"
          }
        },
        "security": [
          {
            "localBearerToken": []
          },
          {
            "localHeaderToken": []
          }
        ],
        "summary": "Clear a scoped local API lockout",
        "x-capability": {
          "operationId": "postv1_management_lockouts_clear",
          "method": "POST",
          "path": "/v1/management/lockouts/clear",
          "maturity": "validated",
          "classification": "mutation",
          "authenticationRequired": true,
          "policyRequired": true,
          "auditRequired": true,
          "scope": "broker-local",
          "providerKinds": [],
          "completionMode": "synchronous",
          "limitationCode": "scoped_lockout_only",
          "reasonCode": "implemented_and_validated",
          "nextAction": "invoke_with_required_controls"
        }
      }
    },
    "/v1/management/secrets": {
      "get": {
        "operationId": "getv1_management_secrets",
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/managedSecretsResponse"
                }
              }
            },
            "description": "Successful response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/managedSecretsResponse"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope"
                    }
                  ]
                }
              }
            },
            "description": "Typed failure response"
          }
        },
        "security": [
          {
            "localBearerToken": []
          },
          {
            "localHeaderToken": []
          }
        ],
        "summary": "List managed secret metadata",
        "x-capability": {
          "operationId": "getv1_management_secrets",
          "method": "GET",
          "path": "/v1/management/secrets",
          "maturity": "read-only",
          "classification": "read",
          "authenticationRequired": true,
          "policyRequired": false,
          "auditRequired": false,
          "scope": "mixed",
          "providerKinds": [
            "local-encrypted-store",
            "vault",
            "openbao",
            "aws-secrets-manager",
            "env",
            "file",
            "exec",
            "bitwarden-bws",
            "onepassword-cli"
          ],
          "completionMode": "synchronous",
          "limitationCode": "metadata_only",
          "reasonCode": "implemented_read_only",
          "nextAction": "read_when_authorized"
        }
      }
    },
...
  },
  "servers": [
    {
      "description": "Development loopback transport",
      "url": "http://127.0.0.1:17890"
    }
  ],
...
}